+++ mi/config.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
+from typing import Any
 
 i = None
 
@@ -11,3 +12,4 @@ def init(**kwargs):
 class Config(BaseModel):
     token: str
     origin_uri: str
+    profile: Any